Refactored movement into module
